refactoring servlets

+++ hw3/src/main/java/ru/itmo/model/Product.java
@@ -2,6 +2,6 @@ package ru.itmo.model;
 
 public record Product(
         String name,
-        int price
+        long price
 ) {
 }
